refactor(ui): split result_table.slint into sub-components and extract UI helpers
